🏷️  Move interfaces
