--- apps/cookbook/src/app/shared/showcase-events/showcase-event.ts
@@ -1,11 +1,0 @@
-export interface ShowcaseEvent {
-  name: string;
-  description?: string;
-  signature?: string;
-}
-
-export interface ShowcaseEventColumns {
-  Name: string;
-  Description?: string;
-  Type?: string;
-}
