--- apps/cookbook/src/app/shared/showcase-methods/showcase-method.ts
@@ -1,11 +1,0 @@
-export interface ShowcaseMethod {
-  name: string;
-  description?: string;
-  signature?: string;
-}
-
-export interface ShowcaseMethodColumns {
-  Name: string;
-  Description?: string;
-  Type?: string;
-}
