+++ apps/cookbook/src/app/shared/showcase-methods/showcase-methods.component.ts
@@ -1,6 +1,16 @@
 import { Component, Input } from '@angular/core';
 
-import { ShowcaseMethod, ShowcaseMethodColumns } from './showcase-method';
+export interface ShowcaseMethod {
+  name: string;
+  description?: string;
+  signature?: string;
+}
+
+export interface ShowcaseMethodColumns {
+  Name: string;
+  Description?: string;
+  Type?: string;
+}
 
 @Component({
   selector: 'cookbook-showcase-methods',
