--- apps/cookbook/src/app/shared/showcase-properties/showcase-property.ts
@@ -1,14 +1,0 @@
-export interface ShowcaseProperty {
-  name: string;
-  description?: string;
-  defaultValue?: string;
-  inputValues?: string[];
-  preserveInputValuesWhitespaces?: boolean;
-}
-
-export interface ShowcasePropertyColumns {
-  Name: string;
-  Description?: string;
-  Type?: string;
-  Default?: string;
-}
