+++ apps/cookbook/src/app/showcase/avatar-showcase/avatar-showcase.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
 
 import { Sizes } from '@kirbydesign/designsystem';
 
-import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-property';
 declare var require: any;
 
 @Component({
