+++ apps/cookbook/src/app/showcase/fab-sheet-showcase/fab-sheet-showcase.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
+import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
 
-import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-property';
 declare var require: any;
 
 @Component({
